+++ swing/common-model/src/main/java/is/codion/swing/common/model/component/combobox/DefaultFilterComboBoxModel.java
@@ -616,7 +616,7 @@ final class DefaultFilterComboBoxModel<T> implements FilterComboBoxModel<T> {
 			}
 
 			@Override
-			public void sortItems() {
+			public void sort() {
 				synchronized (lock) {
 					if (sortInternal()) {
 						notifyChanges();