+++ tools/generator/ui/src/test/java/is/codion/tools/generator/ui/DomainGeneratorPanelTest.java
@@ -34,7 +34,7 @@ public class DomainGeneratorPanelTest {
 		DomainGeneratorModel model = DomainGeneratorModel.domainGeneratorModel(Database.instance(), UNIT_TEST_USER);
 		new DomainGeneratorPanel(model);
 		model.schemaModel().items().refresh();
-		model.schemaModel().items().visible().sortItems();
+		model.schemaModel().items().visible().sort();
 		model.schemaModel().selection().index().set(1);
 		model.populateSelected(schema -> {});
 		model.entityModel().selection().index().set(0);